#473 Added JoinCaptcha Controller, fixed some issues Added response from bot on user joined chat
