+++ modules/bot/components/request/SystemMessageCommandResolver.php
@@ -1,6 +1,7 @@
 <?php
 namespace app\modules\bot\components\request;
 
+use app\modules\bot\controllers\publics\JoinCaptchaController;
 use app\modules\bot\controllers\publics\SystemMessageController;
 use TelegramBot\Api\Types\Update;
 
@@ -9,7 +10,7 @@ class SystemMessageCommandResolver implements ICommandResolver
     public function resolveCommand(Update $update)
     {
         if ($update->getMessage() && $update->getMessage()->getNewChatMember()) {
-            $commandText = SystemMessageController::createRoute('show-captcha', ['joinedUser' => $update->getMessage()->getFrom()->getId()]);
+            $commandText = JoinCaptchaController::createRoute('show-captcha');
         }
 
         if ($update->getMessage() && $update->getMessage()->getLeftChatMember()) {
