+++ modules/bot/controllers/publics/SystemMessageController.php
@@ -27,39 +27,16 @@ class SystemMessageController extends Controller
 
         if (isset($joinHiderStatus) && $joinHiderStatus->value == ChatSetting::JOIN_HIDER_STATUS_ON) {
             $deleteMessage = true;
-
-            $telegramUser = $this->getTelegramUser();
-            BotChatCaptcha::removeCaptchaInfo($chat->id,$telegramUser->provider_user_id);
-
         }
+
+        $telegramUser = $this->getTelegramUser();
+        BotChatCaptcha::removeCaptchaInfo($chat->id,$telegramUser->provider_user_id);
 
         if ($deleteMessage) {
             return $this->getResponseBuilder()
             ->deleteMessage()
             ->build();
         }
-    }
-
-    public function actionShowCaptcha()
-    {
-        $chat = $this->getTelegramChat();
-        $joinCaptchaStatus = $chat->getSetting(ChatSetting::JOIN_CAPTCHA_STATUS);
-
-        if (isset($joinCaptchaStatus) && $joinCaptchaStatus->value == ChatSetting::JOIN_CAPTCHA_STATUS_ON) {
-
-            $telegramUser = $this->getTelegramUser();
-
-            $needShowCaptcha = BotChatCaptcha::checkCaptcha($chat->id,$telegramUser->provider_user_id);
-
-            if ($needShowCaptcha) {
-
-                $mesageText = new MessageText('Here will be a captcha action');
-                return $this->getResponseBuilder()
-                    ->sendMessage($mesageText)
-                    ->build();
-            }
-        }
-
     }
 
     public function actionGroupToSupergroup()
