+++ modules/bot/views/publics/join-captcha/show-captcha.php
@@ -1,0 +1,9 @@
+<?php
+/**
+ * @var $chatName string
+ * @var $firstName string
+ * @var $lastName string
+ */
+?>
+<?= Yii::t('bot', 'Welcome to group') . ' ' . $chatName . ',' ?><br/>
+<?= $firstName . ' ' . $lastName ?>
